- Changed stars behavior on gameplay background: now they move top to bottom like the player is moving through space.

+++ Assets/Scripts/Misc/RandomBetweenTwoConst.cs
@@ -20,6 +20,6 @@ public class RandomBetweenTwoConst
 
 	public float Random()
 	{
-		return UnityEngine.Random.Range(min, max);
+		return UnityEngine.Random.Range((float)min, (float)max);
 	}
 }
